+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Migrated `golangci-lint` to version `v2.1.6`
+
 ## v3.111.0
 * Added `sugar.PrintErrorWithoutStack` helper for remove stack records from error string
 * Added `sugar.UnwrapError` helper for unwrap source error to root errors